+++ Gateway/src/main/java/at/fhjoanneum/ippr/gateway/security/persistence/objects/Organization.java
@@ -1,16 +1,24 @@
 package at.fhjoanneum.ippr.gateway.security.persistence.objects;
 
+import at.fhjoanneum.ippr.gateway.security.persistence.entities.UserImpl;
+
+import java.util.List;
+
 public interface Organization {
 
     Long getOId();
 
     String getSystemId();
 
-    String getOrganisationName();
+    String getOrganizationName();
 
     void setOrganizationName(String organizationName);
 
     String getDescription();
 
     void setOrganizationDescription(String organizationDescription);
+
+    List<UserImpl> getEmployees();
+
+    void setEmployees(List<UserImpl> employees);
 }