+++ Gateway/src/main/java/at/fhjoanneum/ippr/gateway/security/rbacmapping/retrieval/RBACRetrievalService.java
@@ -6,5 +6,5 @@ import at.fhjoanneum.ippr.gateway.security.persistence.entities.cache.CacheUser;
 
 public interface RBACRetrievalService {
 
-  Map<String, CacheUser> getSystemUsers();
+  Map<String, CacheUser> getSystemUsers(String path);
 }