+++ Gateway/src/main/java/at/fhjoanneum/ippr/gateway/security/services/RBACService.java
@@ -1,6 +1,7 @@
 package at.fhjoanneum.ippr.gateway.security.services;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.concurrent.Future;
 
 import at.fhjoanneum.ippr.commons.dto.user.UserDTO;
@@ -14,4 +15,7 @@ public interface RBACService {
   public Future<List<UserDTO>> getUsersOfRule(List<String> ruleNames);
 
   public Future<List<Rule>> getRules();
+
+  public Optional<User> updateUser(Long uId, String username, String firstName, String lastName, String email,
+                                   String password, Long o_id);
 }